Create `BackendClient+Requests.swift`
