+++ Mlem/Packages/MlemBackend/Sources/MlemBackend/BackendClient.swift
@@ -7,6 +7,7 @@
 
 import Foundation
 import os
+import Rest
 import SwiftUI
 import MlemLogger
 
@@ -24,9 +25,12 @@ public enum BackendEnvironment {
 @Observable
 public class BackendClient {
     let log: Logger = .mlemLogger()
+
+    internal let restClient = RestClient()
     
-    public private(set) var environment: BackendEnvironment = .production
-    private let jsonDecoder: JSONDecoder = {
+    public internal(set) var environment: BackendEnvironment = .production
+
+    internal let jsonDecoder: JSONDecoder = {
         let decoder: JSONDecoder = .init()
         decoder.dateDecodingStrategy = .custom { decoder in
             let formatter: ISO8601DateFormatter = .init()
@@ -40,31 +44,14 @@ public class BackendClient {
         return decoder
     }()
     
-    public private(set) var flairs: MlemFlairs = .init(developers: .init())
-    public private(set) var testflightUpdate: URL?
+    public internal(set) var flairs: MlemFlairs = .init(developers: .init())
+    public internal(set) var testflightUpdate: URL?
     
     public static var main: BackendClient = .init()
-    private var baseUrl: URL { environment.address }
+    internal var baseUrl: URL { environment.address }
     
-    private init() {
+    internal init() {
         refresh()
-    }
-    
-    public func healthCheck() async throws -> BackendHealthCheck {
-        let (data, _) = try await URLSession.shared.data(for: URLRequest(url: baseUrl.appendingPathComponent("/v0/health")))
-        return try jsonDecoder.decode(BackendHealthCheck.self, from: data)
-    }
-    
-    public func getInstances() async throws -> [InstanceSummary] {
-        let request: URLRequest = .init(url: baseUrl
-            .appendingPathComponent("/v1/stats/instances")
-            .appending(queryItems: [
-                .init(name: "minTotalUsers", value: "20"),
-                .init(name: "minMonthyUsers", value: "1")
-            ])
-        )
-        let (data, _) = try await URLSession.shared.data(for: request)
-        return try jsonDecoder.decode([InstanceSummary].self, from: data)
     }
     
     public func changeEnvironment(to environment: BackendEnvironment) {
@@ -72,7 +59,7 @@ public class BackendClient {
         refresh()
     }
     
-    private func refresh() {
+    internal func refresh() {
         Task {
             do {
                 try await fetchFlairs()
@@ -81,27 +68,5 @@ public class BackendClient {
                 log.error("\(error.localizedDescription)")
             }
         }
-    }
-    
-    private func fetchTestflightUpdate() async throws {
-        let (data, _) = try await URLSession.shared.data(for: URLRequest(url: baseUrl.appendingPathComponent("/v0/mlem/testflight")))
-        testflightUpdate = try jsonDecoder.decode(TestflightUpdate.self, from: data).url
-    }
-    
-    private func fetchFlairs(enabledOnly: Bool = true) async throws {
-        let request: URLRequest = .init(url: baseUrl
-            .appendingPathComponent("/v0/mlem/flairs")
-            .appending(queryItems: [
-                .init(name: "enabledOnly", value: enabledOnly.description)
-            ])
-        )
-        let (data, _) = try await URLSession.shared.data(for: request)
-        let response = try jsonDecoder.decode([MlemFlair].self, from: data)
-        
-        flairs = .init(developers: .init(
-            response
-                .filter { [.activeDev, .inactiveDev].contains($0.flairType) }
-                .map(\.apId)
-        ))
     }
 }
